[GUI APi] Move widgets to new file, and renaming gui_componenet.h to comonenet.h (Spelling bad)
