+++ src/client/client.cpp
@@ -3,7 +3,7 @@
 #include "client_config.h"
 #include "gl/gl_errors.h"
 #include "gl/primitive.h"
-#include "gui/overlay_component.h"
+#include "gui/component.h"
 #include "input/keyboard.h"
 #include "world/chunk_mesh_generation.h"
 #include <SFML/Window/Window.hpp>
