+++ src/client/gui/component.h
